+++ backend/src/api/endpoints/hls_api.rs
@@ -7,7 +7,7 @@ use crate::api::model::UserSession;
 use crate::api::model::{create_custom_video_stream_response, CustomVideoStreamType};
 use crate::api::model::AppState;
 use crate::auth::Fingerprint;
-use crate::model::ProxyUserCredentials;
+use crate::model::{ConfigTarget, ProxyUserCredentials};
 use crate::model::{ConfigInput, InputSource};
 use crate::processing::parser::hls::{
     get_hls_session_token_and_url_from_token, rewrite_hls, RewriteHlsProps,
@@ -17,9 +17,11 @@ use axum::http::HeaderMap;
 use axum::response::IntoResponse;
 use log::{debug, error};
 use serde::Deserialize;
-use shared::model::{PlaylistItemType, UserConnectionPermission, XtreamCluster};
+use shared::model::{PlaylistItemType, StreamChannel, TargetType, UserConnectionPermission, XtreamCluster};
 use shared::utils::{is_hls_url, replace_url_extension, sanitize_sensitive_info, CUSTOM_VIDEO_PREFIX, HLS_EXT};
 use std::sync::Arc;
+use crate::repository::m3u_repository::m3u_get_item_for_stream_id;
+use crate::repository::xtream_repository;
 
 #[derive(Debug, Deserialize)]
 struct HlsApiPathParams {
@@ -151,6 +153,49 @@ pub(in crate::api) async fn handle_hls_stream_request(
     }
 }
 
+async fn get_stream_channel(app_state: &Arc<AppState>, target: &Arc<ConfigTarget>, virtual_id: u32) -> Option<StreamChannel> {
+    if target.has_output(TargetType::Xtream) {
+        if let Ok((pli, _mapping)) = xtream_repository::xtream_get_item_for_stream_id(
+            virtual_id,
+            app_state,
+            target,
+            None
+        ).await {
+            Some(pli.to_stream_channel())
+        } else {
+            None
+        }
+    } else {
+        match m3u_get_item_for_stream_id(virtual_id, app_state, target).await {
+            Ok(pli) => Some(pli.to_stream_channel()),
+            Err(_) => { None }
+        }
+    }
+}
+
+async fn resolve_stream_channel(
+    app_state: &Arc<AppState>,
+    target: &Arc<ConfigTarget>,
+    virtual_id: u32,
+    hls_url: &str,
+) -> StreamChannel {
+    let mut channel = match get_stream_channel(app_state, target, virtual_id).await {
+        Some(channel) => channel,
+        None => StreamChannel {
+            virtual_id,
+            provider_id: 0,
+            item_type: PlaylistItemType::LiveHls,
+            cluster: XtreamCluster::Live,
+            group: "Unknown".to_string(),
+            title: "Unknown".to_string(),
+            url: hls_url.to_string(),
+        },
+    };
+
+    channel.item_type = PlaylistItemType::LiveHls;
+    channel
+}
+
 #[allow(clippy::too_many_lines)]
 async fn hls_api_stream(
     Fingerprint(fingerprint, addr): Fingerprint,
@@ -207,7 +252,7 @@ async fn hls_api_stream(
                 &app_state.app_config,
                 CustomVideoStreamType::ProviderConnectionsExhausted,
             )
-                .into_response();
+            .into_response();
         }
 
         let hls_url = match get_hls_session_token_and_url_from_token(
@@ -218,21 +263,22 @@ async fn hls_api_stream(
             _ => return axum::http::StatusCode::BAD_REQUEST.into_response(),
         };
 
-        session.stream_url = hls_url;
+        session.stream_url.clone_from(&hls_url);
         if session.virtual_id == virtual_id {
             if is_seek_request(XtreamCluster::Live, &req_headers).await {
                 // partial request means we are in reverse proxy mode, seek happened
+                let stream_channel = resolve_stream_channel(&app_state, &target, virtual_id, &hls_url).await;
                 return force_provider_stream_response(
                     &addr,
                     &app_state,
                     session,
-                    PlaylistItemType::LiveHls,
+                    stream_channel,
                     &req_headers,
                     &input,
                     &user,
                 )
-                    .await
-                    .into_response();
+                .await
+                .into_response();
             }
         } else {
             return axum::http::StatusCode::BAD_REQUEST.into_response();
@@ -264,11 +310,13 @@ async fn hls_api_stream(
                 .into_response();
         }
 
+        let stream_channel = resolve_stream_channel(&app_state, &target, virtual_id, &hls_url).await;
+
         force_provider_stream_response(
             &addr,
             &app_state,
             session,
-            PlaylistItemType::LiveHls,
+            stream_channel,
             &req_headers,
             &input,
             &user,
@@ -279,6 +327,7 @@ async fn hls_api_stream(
         axum::http::StatusCode::BAD_REQUEST.into_response()
     }
 }
+
 
 pub fn hls_api_register() -> axum::Router<Arc<AppState>> {
     axum::Router::new().route(