+++ shared/src/model/web_socket.rs
@@ -1,6 +1,6 @@
 use std::io;
 use bytes::Bytes;
-use crate::model::{ConfigType, PlaylistUpdateState, StatusCheck};
+use crate::model::{ActiveUserConnectionChange, ConfigType, PlaylistUpdateState, StatusCheck};
 use serde::{Deserialize, Serialize};
 
 pub const PROTOCOL_VERSION: u8 = 1;
@@ -79,7 +79,7 @@ pub enum ProtocolMessage {
     ServerError(String),
     StatusRequest(String),
     StatusResponse(StatusCheck),
-    ActiveUserResponse(usize, usize), // user_count, connection count
+    ActiveUserResponse(ActiveUserConnectionChange),
     ActiveProviderResponse(String, usize),
     ConfigChangeResponse(ConfigType),
     PlaylistUpdateResponse(PlaylistUpdateState),